+++ src/bitmap/color_plate_scanner.cpp
@@ -88,7 +88,7 @@ namespace Invader {
 
     #define GET_PIXEL(x,y) (pixels[y * width + x])
 
-    GeneratedBitmapData ColorPlateScanner::scan_color_plate(const ColorPlatePixel *pixels, std::uint32_t width, std::uint32_t height, BitmapType type, BitmapUsage usage, float bump_height, const std::optional<ColorPlateScannerSpriteParameters> &sprite_parameters, std::int16_t mipmaps, ScannedColorMipmapType mipmap_type, std::optional<float> mipmap_fade_factor) {
+    GeneratedBitmapData ColorPlateScanner::scan_color_plate(const ColorPlatePixel *pixels, std::uint32_t width, std::uint32_t height, BitmapType type, BitmapUsage usage, float bump_height, const std::optional<ColorPlateScannerSpriteParameters> &sprite_parameters, std::int16_t mipmaps, ScannedColorMipmapType mipmap_type, std::optional<float> mipmap_fade_factor, std::optional<float> sharpen, std::optional<float> blur) {
         ColorPlateScanner scanner;
         GeneratedBitmapData generated_bitmap;
 
@@ -282,7 +282,7 @@ namespace Invader {
 
         // If we aren't making interface bitmaps, generate mipmaps when needed
         if(type != BitmapType::BITMAP_TYPE_INTERFACE_BITMAPS) {
-            generate_mipmaps(generated_bitmap, mipmaps, mipmap_type, mipmap_fade_factor, sprite_parameters);
+            generate_mipmaps(generated_bitmap, mipmaps, mipmap_type, mipmap_fade_factor, sprite_parameters, sharpen, blur);
         }
 
         // If we're making cubemaps, we need to make all sides of each cubemap sequence one cubemap bitmap data. 3D textures work similarly
@@ -581,7 +581,7 @@ namespace Invader {
         }
     }
 
-    void ColorPlateScanner::generate_mipmaps(GeneratedBitmapData &generated_bitmap, std::int16_t mipmaps, ScannedColorMipmapType mipmap_type, std::optional<float> mipmap_fade_factor, const std::optional<ColorPlateScannerSpriteParameters> &sprite_parameters) {
+    void ColorPlateScanner::generate_mipmaps(GeneratedBitmapData &generated_bitmap, std::int16_t mipmaps, ScannedColorMipmapType mipmap_type, std::optional<float> mipmap_fade_factor, const std::optional<ColorPlateScannerSpriteParameters> &sprite_parameters, std::optional<float> sharpen, std::optional<float> blur) {
         auto mipmaps_unsigned = static_cast<std::uint32_t>(mipmaps);
         for(auto &bitmap : generated_bitmap.bitmaps) {
             std::uint32_t mipmap_width = bitmap.width;
@@ -689,6 +689,45 @@ namespace Invader {
                             else {
                                 pixel.alpha = UINT8_MAX;
                             }
+                        }
+                    }
+                }
+
+                // Apply a sharpen filter? https://en.wikipedia.org/wiki/Unsharp_masking
+                if(sharpen.has_value() && sharpen.value() > 0.0F) {
+                    const float &SHARPEN_VALUE = sharpen.value();
+
+                    // Make a copy of the mipmap to work off of
+                    std::vector<ColorPlatePixel> unsharpened_pixels(this_mipmap_data, this_mipmap_data + mipmap_width * mipmap_height);
+
+                    // Go through each pixel and apply the sharpening filter
+                    for(std::uint32_t y = 0; y < mipmap_height; y++) {
+                        for(std::uint32_t x = 0; x < mipmap_width; x++) {
+                            auto &center = unsharpened_pixels[x + y * mipmap_width];
+                            auto &left = (x == 0) ? center : unsharpened_pixels[x + y * mipmap_width - 1];
+                            auto &right = (x + 1 == mipmap_width) ? center : unsharpened_pixels[x + y * mipmap_width + 1];
+                            auto &top = (y == 0) ? center : unsharpened_pixels[x + (y - 1) * mipmap_width];
+                            auto &bottom = (y + 1 == mipmap_height) ? center : unsharpened_pixels[x + (y + 1) * mipmap_width];
+                            auto &this_pixel = this_mipmap_data[x + y * mipmap_width];
+
+                            #define APPLY_SHARPEN(channel) { \
+                                std::int32_t modification = static_cast<std::int32_t>(center.channel) * (1.0 + 4.0F * SHARPEN_VALUE) - (static_cast<std::int32_t>(top.channel) + left.channel + bottom.channel + right.channel) * SHARPEN_VALUE; \
+                                if(modification > 0xFF) { \
+                                    this_pixel.channel = 0xFF; \
+                                } \
+                                else if(modification < 0x00) { \
+                                    this_pixel.channel = 0x00; \
+                                } \
+                                else { \
+                                    this_pixel.channel = static_cast<std::uint8_t>(modification); \
+                                } \
+                            }
+
+                            APPLY_SHARPEN(red);
+                            APPLY_SHARPEN(green);
+                            APPLY_SHARPEN(blue);
+
+                            #undef APPLY_SHARPEN
                         }
                     }
                 }